+++ modules/core/client-react/entry.tsx
